imhotep_app/requirements.txt
PyQt5
qtawesome
mysql-connector-python
python-dotenv


User_ID *
Patient_ID *